services/retriever_drafter/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-bigquery==3.25.0
google-cloud-aiplatform==1.70.0
pydantic==2.9.2
